feat: adiciona mensagem de validacao para tentativa de remocao de produto associado a uma categoria

+++ src/main/java/com/devcaiqueoliveira/mercatopdvsystem/service/CategoryService.java
@@ -4,6 +4,7 @@ import com.devcaiqueoliveira.mercatopdvsystem.entity.Category;
 import com.devcaiqueoliveira.mercatopdvsystem.exception.BusinessRuleException;
 import com.devcaiqueoliveira.mercatopdvsystem.exception.EntityNotFoundException;
 import com.devcaiqueoliveira.mercatopdvsystem.repository.CategoryRepository;
+import com.devcaiqueoliveira.mercatopdvsystem.repository.ProductRepository;
 import com.devcaiqueoliveira.mercatopdvsystem.service.validator.category.CategoryValidatorStrategy;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,6 +19,7 @@ public class CategoryService {
 
     private final CategoryRepository repository;
     private final List<CategoryValidatorStrategy> validators;
+    private final ProductRepository productRepository;
 
     public Category findById(Long id) {
         return repository.findById(id)
@@ -45,6 +47,11 @@ public class CategoryService {
         if (!repository.existsById(id)) {
             throw new EntityNotFoundException("Categoria não encontrada para ser removida");
         }
+
+        if (productRepository.existsByCategory(id)) {
+            throw new BusinessRuleException("Não é possível excluir a categoria pois existem produtos vinculados a ela.");
+        }
+
         repository.deleteById(id);
     }
